refactor: capacitor 딥링크 리스너 로직 수정(#288)
